Adds `invoice_number` to the `Order#to_hash` return value
I found that the Invoice Number wasn't being updated on the server
when calling `CIM::Transaction#create_transaction_auth_only` with
and `Order` object. Upon reviewing the documentation, the correct key
for the Invoice Number is `invoice_number`, not `invoice_num`.

This commit keeps `invoice_num` and creates an `invoice_number` alias to preserve backwards compatability.