Add solution id to the AIM transaction objects
Solution id provides the functionality to review transactions on the basis
of single id for a partner devloper. So adding an option to provide the
solution_id when the transaction is created in old model for AIM
transactions. Also adding a test case to check the solution_id is set
properly from the response returned from the server.

Signed-off-by: Srijan Misra <srmisra@WW975L29177B.visa.com>